requirements.txt
boto==2.25.0
distribute==0.6.31
feedparser==5.1.3
Flask-S3==0.1.6
Flask==0.10.1
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
Jinja2==2.7.2
MarkupSafe==0.18
oauth2==1.5.211
oauthlib==0.6.0
pymongo==2.6.3
python-dateutil==2.2
pytz==2013.9
PyYAML==3.10
readability-api==0.2.3
redis==2.9.1
requests-oauthlib==0.4.0
requests==2.2.1
rq==0.3.13
simplejson==3.3.2
six==1.5.2
times==0.6.2
webassets==0.9
Werkzeug==0.9.4
wsgiref==0.1.2